chore(lints): clear the clippy remainder so the gate can be turned on
1063 -> 0. Most of it was mechanical (clippy --fix), reviewed by category
before trusting; the rest hand-fixed. Notes on the parts that were not
mechanical:

- input_replay dropped a full-bundle memcpy per iteration: both arms of a
  speed check cloned the same bundle, and the source was never read again.
- st2110/video had `if pending.is_empty() { 6 } else { 6 }` under a comment
  describing an up-front reservation the code never performs. Every triplet
  is charged 6 at its own iteration, which is self-consistent, so the
  conditional was vestigial. Behaviour and the wire format are unchanged.
- classify_kms_error matched display_device_invalid explicitly and again in
  the fallback; both reported the same code, so the arm gained nothing.
- Timecode::to_string became a Display impl, so to_string() comes free.
- pes_splice's LATM writer builds its bit vector by literal now.

Four `unwrap` after `is_some` sites became edition-2024 let-chains.

Also fixed one cargo-fix casualty: derivable_impls replaced two manual
Default impls with derives, which orphaned two `new()` constructors that
only tests called. They were exactly Default, so they are gone and the
callers use default().

cargo clippy --all-targets -- -D warnings passes. 1912 tests pass.

+++ src/engine/input_rtmp.rs
@@ -260,7 +260,7 @@ async fn process_media(
                             1 => {
                                 // AVC NALU data
                                 let is_keyframe = frame_type == 1;
-                                let composition_time = ((data[2] as i32) << 16 | (data[3] as i32) << 8 | data[4] as i32) as i32;
+                                let composition_time = (data[2] as i32) << 16 | (data[3] as i32) << 8 | data[4] as i32;
                                 // Sign extend 24-bit
                                 let composition_time = if composition_time & 0x800000 != 0 {
                                     composition_time | !0xFFFFFF
